Validation
Added validation framework for returning global and field level errors from REST services. Also added Springfox API which will expose endpoints via Swagger. This does not work for Spring Data REST as there is a known defect. We will have to wait for a newer release for that to work. Spring Data REST endpoints are exposed via the HAL Browser.