Search Queries
Added queries and tests to support searching for shifts at a courthouse and sheriffs at a courthouse. Added some framework for Dates and Unit tests to make it simplier to create objects. There is some weirdness around dates since we are currently storing timezones. We may need to remove that.