Endpoint Updates
I added a shift tab to the excel worksheet so we can quickly add new records. I also removed all the @JsonIgnore directives so the hyperlinking via HATEOAS works in the REST API.
Search Shifts

Fixed a bug in the search for shifts based on a date range. Used a native query instead of HQL since postgres was having issues changing date to timestamp.