Get Duties for a Courthouse
Added method for getting duties for a given courthouse and date range. Fixed a bug with SimpleDuty where it was not returning the idPath. Moved some properties up a level to stop unit tests behaving differently that application regarding JSON dates. Added library so that PATCH will work in JUnit tests. Changed delete endpoints to patch for expiring assignment and dutyrecurrence.